+++ src/knowledge/dialogs/simple-prompt/SimplePromptInterfaceTools.ts
@@ -1,8 +1,7 @@
 import { spaceTrim } from 'spacetrim';
 import { PipelineExecutionError } from '../../../errors/PipelineExecutionError';
 import type { CommonExecutionToolsOptions } from '../../../execution/CommonExecutionToolsOptions';
-import type { UserInterfaceTools } from '../../../execution/UserInterfaceTools';
-import type { UserInterfaceToolsPromptDialogOptions } from '../../../execution/UserInterfaceTools';
+import type { UserInterfaceTools, UserInterfaceToolsPromptDialogOptions } from '../../../execution/UserInterfaceTools';
 
 /**
  * Wrapper around `window.prompt` synchronous function that interacts with the user via browser prompt
@@ -16,7 +15,7 @@ export class SimplePromptInterfaceTools implements UserInterfaceTools {
     public constructor(private readonly options: CommonExecutionToolsOptions = {}) {}
 
     /**
-     * Trigger window.PROMPT DIALOG
+     * Trigger window.DIALOG BLOCK
      */
     public async promptDialog(options: UserInterfaceToolsPromptDialogOptions): Promise<string> {
         const answer = window.prompt(